Lwm2m: change start lhServer/bootstrapserver
