+++ common/transport/lwm2m/src/main/java/org/thingsboard/server/transport/lwm2m/server/LwM2mTransportServerConfiguration.java
@@ -29,14 +29,17 @@ import org.eclipse.leshan.server.model.LwM2mModelProvider;
 import org.eclipse.leshan.server.security.DefaultAuthorizer;
 import org.eclipse.leshan.server.security.EditableSecurityStore;
 import org.eclipse.leshan.server.security.SecurityChecker;
-import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Component;
 import org.thingsboard.server.common.data.StringUtils;
 import org.thingsboard.server.queue.util.TbLwM2mTransportComponent;
 import org.thingsboard.server.transport.lwm2m.config.LwM2MTransportServerConfig;
+import org.thingsboard.server.transport.lwm2m.secure.LWM2MGenerationPSkRPkECC;
 import org.thingsboard.server.transport.lwm2m.server.client.LwM2mClientContext;
 import org.thingsboard.server.transport.lwm2m.utils.LwM2mValueConverterImpl;
 
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
 import java.math.BigInteger;
 import java.security.AlgorithmParameters;
 import java.security.KeyFactory;
@@ -70,25 +73,54 @@ public class LwM2mTransportServerConfiguration {
     private PublicKey publicKey;
     private PrivateKey privateKey;
     private boolean pskMode = false;
+    private LeshanServer leshanServer;
     private final LwM2mTransportContextServer context;
     private final CaliforniumRegistrationStore registrationStore;
     private final EditableSecurityStore securityStore;
     private final LwM2mClientContext lwM2mClientContext;
+    private final LwM2mTransportServiceImpl service;
 
-    public LwM2mTransportServerConfiguration(LwM2mTransportContextServer context, CaliforniumRegistrationStore registrationStore, EditableSecurityStore securityStore, LwM2mClientContext lwM2mClientContext) {
+    public LwM2mTransportServerConfiguration(LwM2mTransportContextServer context, CaliforniumRegistrationStore registrationStore,
+                                             EditableSecurityStore securityStore, LwM2mClientContext lwM2mClientContext,
+                                             @Lazy LwM2mTransportServiceImpl service) {
         this.context = context;
         this.registrationStore = registrationStore;
         this.securityStore = securityStore;
         this.lwM2mClientContext = lwM2mClientContext;
+        this.service = service;
     }
 
-    @Bean
+    @PostConstruct
+    public void init() {
+        if (this.context.getLwM2MTransportServerConfig().getEnableGenNewKeyPskRpk()) {
+            new LWM2MGenerationPSkRPkECC();
+        }
+        this.leshanServer = this.getLhServer();
+        this.startLhServer();
+    }
+
+    private void startLhServer() {
+        this.leshanServer.start();
+        LwM2mServerListener lhServerCertListener = new LwM2mServerListener(service);
+        this.leshanServer.getRegistrationService().addListener(lhServerCertListener.registrationListener);
+        this.leshanServer.getPresenceService().addListener(lhServerCertListener.presenceListener);
+        this.leshanServer.getObservationService().addListener(lhServerCertListener.observationListener);
+    }
+
     public LeshanServer getLeshanServer() {
-        log.info("Starting LwM2M transport Server... PostConstruct");
-        return this.getLhServer(this.context.getLwM2MTransportServerConfig().getPort(), this.context.getLwM2MTransportServerConfig().getSecurePort());
+        return  this.leshanServer;
     }
 
-    private LeshanServer getLhServer(Integer serverPortNoSec, Integer serverSecurePort) {
+    @PreDestroy
+    public void shutdown() {
+        log.info("Stopping LwM2M transport Server!");
+        leshanServer.destroy();
+        log.info("LwM2M transport Server stopped!");
+    }
+
+    private LeshanServer getLhServer() {
+        Integer serverPortNoSec = this.context.getLwM2MTransportServerConfig().getPort();
+        Integer serverSecurePort = this.context.getLwM2MTransportServerConfig().getSecurePort();
         LeshanServerBuilder builder = new LeshanServerBuilder();
         builder.setLocalAddress(this.context.getLwM2MTransportServerConfig().getHost(), serverPortNoSec);
         builder.setLocalSecureAddress(this.context.getLwM2MTransportServerConfig().getSecureHost(), serverSecurePort);
